Changed style of code... now all single line conditionals/loops are in the same line as the condition itself

+++ include/StringFuncs.hpp
@@ -15,7 +15,7 @@ namespace Str
 
 	std::string ToUpper( const std::string & str );
 
-	int FetchArgs( const std::vector< std::string > & args, std::map< std::string, std::string > & argmap, int from );
+	int FetchArgs( const std::vector< std::string > & args, std::map< std::string, std::string > & argmap, size_t from );
 
 	bool Replace( std::string & str, const std::string & from, const std::string & to );
 }
